Update processMarkedLines to execute regex on correct variable and improve regex
The RegExp was improved to avoid ReDos attacks. See the following resources for more information:

https://www.regular-expressions.info/catastrophic.html
https://github.blog/2023-05-09-how-to-fix-a-redos

This (tool)[https://devina.io/redos-checker] was used to fuzz both the old and new regexp and shows that the old regexp was vulernable whereas the new one is not.

We also fix a var reference error within preprocess() which resulted in an extra newline always being added before line splitters.